Ref: Created /user_transactions_tokens and moved files inside
