+++ app/src/components/market/sections/market_bond/market_bond.tsx
@@ -22,15 +22,15 @@ import { ButtonType } from '../../../button/button_styling_types'
 import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
 import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
-import { AssetBalance } from '../../common/asset_balance'
 import { CurrenciesWrapper, GenericError } from '../../common/common_styled'
-import { TransactionDetailsCard } from '../../common/transaction_details_card'
-import { TransactionDetailsLine } from '../../common/transaction_details_line'
-import { TransactionDetailsRow, ValueStates } from '../../common/transaction_details_row'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
 import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
 import { MarketScale } from '../common_sections/card_bottom_details/market_scale'
 import { OutcomeTable } from '../common_sections/tables/outcome_table'
+import { AssetBalance } from '../common_sections/user_transactions_tokens/asset_balance'
+import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction_details_card'
+import { TransactionDetailsLine } from '../common_sections/user_transactions_tokens/transaction_details_line'
+import { TransactionDetailsRow, ValueStates } from '../common_sections/user_transactions_tokens/transaction_details_row'
 
 interface Props extends RouteComponentProps<any> {
   marketMakerData: MarketMakerData
