+++ app/src/components/market/sections/market_create/steps/items/funding_and_fee_step.tsx
@@ -39,11 +39,6 @@ import {
   OutcomesTable,
   OutcomesTableWrapper,
 } from '../../../../common/common_styled'
-import { CurrencySelector } from '../../../../common/currency_selector'
-import { TradingFeeSelector } from '../../../../common/trading_fee_selector'
-import { TransactionDetailsCard } from '../../../../common/transaction_details_card'
-import { TransactionDetailsLine } from '../../../../common/transaction_details_line'
-import { TransactionDetailsRow, ValueStates } from '../../../../common/transaction_details_row'
 import { SetAllowance } from '../../../common_sections/allowance/set_allowance'
 import { DisplayArbitrator } from '../../../common_sections/arbitrators/display_arbitrator'
 import { GridTransactionDetails } from '../../../common_sections/card_bottom_details/grid_transaction_details'
@@ -51,6 +46,14 @@ import { MarketScale } from '../../../common_sections/card_bottom_details/market
 import { CreateCard } from '../../../common_sections/create_card'
 import { VerifiedRow } from '../../../common_sections/create_card/verified_row'
 import { WarningMessage } from '../../../common_sections/message_text/warning_message'
+import { CurrencySelector } from '../../../common_sections/user_transactions_tokens/currency_selector'
+import { TradingFeeSelector } from '../../../common_sections/user_transactions_tokens/trading_fee_selector'
+import { TransactionDetailsCard } from '../../../common_sections/user_transactions_tokens/transaction_details_card'
+import { TransactionDetailsLine } from '../../../common_sections/user_transactions_tokens/transaction_details_line'
+import {
+  TransactionDetailsRow,
+  ValueStates,
+} from '../../../common_sections/user_transactions_tokens/transaction_details_row'
 import { Outcome } from '../outcomes'
 
 const CreateCardTop = styled(CreateCard)`
