+++ app/src/components/market/sections/market_pooling/market_pool_liquidity.tsx
@@ -22,15 +22,15 @@ import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
 import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
-import { CurrencySelector } from '../../common/currency_selector'
-import { TokenBalance } from '../../common/token_balance'
-import { TransactionDetailsCard } from '../../common/transaction_details_card'
-import { TransactionDetailsLine } from '../../common/transaction_details_line'
-import { TransactionDetailsRow, ValueStates } from '../../common/transaction_details_row'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
 import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
 import { WarningMessage } from '../common_sections/message_text/warning_message'
 import { OutcomeTable } from '../common_sections/tables/outcome_table'
+import { CurrencySelector } from '../common_sections/user_transactions_tokens/currency_selector'
+import { TokenBalance } from '../common_sections/user_transactions_tokens/token_balance'
+import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction_details_card'
+import { TransactionDetailsLine } from '../common_sections/user_transactions_tokens/transaction_details_line'
+import { TransactionDetailsRow, ValueStates } from '../common_sections/user_transactions_tokens/transaction_details_row'
 
 import { UserPoolData } from './user_pool_data'
